Wraper up python building script. tracker item 2019
git-svn-id: https://opencmiss.svn.sourceforge.net/svnroot/opencmiss/cm/trunk@754 61e36070-c886-4827-b7c0-ed9e4dc59cc8

+++ examplesBuilder.py
@@ -9,25 +9,28 @@ if not os.path.isdir(logDir):
   os.mkdir(cwd + "/../../build")
   os.mkdir(logDir);
 compiler = sys.argv[1];
-f = open(logDir+'/failedBuilds',"w")
+f = open(logDir+'/successBuilds',"w")
 
 def buildExample(path) :
-   global compiler,logDir;
-   newDir = logDir
-   for folder in path.split('/') :
-     newDir = newDir + '/' + folder
-     if not os.path.isdir(newDir):
-       os.mkdir(newDir)
-   os.chdir(path)
-   if os.path.exists(newDir + "/build-" + compiler) :
-     os.remove(newDir + "/build-" + compiler)
-   err=os.system("make COMPILER=" + compiler + " > " + newDir + "/build-" + compiler +" 2>&1")
-   if err==0 :
-     print "Building %s: <a class='success' href='%slogs_x86_64-linux/%s/build-%s'>success</a><br>" %(path,rootUrl,path,compiler)
+   global compiler,logDir,libSuccess;
+   if (libSuccess!=-1) :
+     newDir = logDir
+     for folder in path.split('/') :
+       newDir = newDir + '/' + folder
+       if not os.path.isdir(newDir):
+         os.mkdir(newDir)
+     os.chdir(path)
+     if os.path.exists(newDir + "/build-" + compiler) :
+       os.remove(newDir + "/build-" + compiler)
+     err=os.system("make COMPILER=" + compiler + " > " + newDir + "/build-" + compiler +" 2>&1")
+     if err==0 :
+       f.write(path+'\n')
+       print "Building %s: <a class='success' href='%slogs_x86_64-linux/%s/build-%s'>success</a><br>" %(path,rootUrl,path,compiler)
+     else :
+       print "Building %s: <a class='fail' href='%slogs_x86_64-linux/%s/build-%s'>failed</a><br>" %(path,rootUrl,path,compiler)
+     os.chdir(cwd)
    else :
-     f.write(path+'\n')
-     print "Building %s: <a class='fail' href='%slogs_x86_64-linux/%s/build-%s'>failed</a><br>" %(path,rootUrl,path,compiler)
-   os.chdir(cwd)
+     print "Building %s: <a class='fail'>failed</a> due to library build failure<br>" %(path)
    return;
 
 def buildLibrary() :
@@ -49,50 +52,49 @@ def buildLibrary() :
    
 libSuccess = buildLibrary()
 
-if libSuccess==0 :
-  #buildExample("ClassicalField/AnalyticLaplace")
-  #buildExample("ClassicalField/Diffusion")
-  #buildExample("ClassicalField/Helmholtz")
-  #buildExample("ClassicalField/Laplace")
-  #buildExample("ClassicalField/NonlinearPoisson")
-  buildExample("ClassicalField/NewLaplace")
-  buildExample("ClassicalField/NumberLaplace")
+#buildExample("ClassicalField/AnalyticLaplace")
+#buildExample("ClassicalField/Diffusion")
+#buildExample("ClassicalField/Helmholtz")
+#buildExample("ClassicalField/Laplace")
+#buildExample("ClassicalField/NonlinearPoisson")
+buildExample("ClassicalField/NewLaplace")
+buildExample("ClassicalField/NumberLaplace")
 
-  #buildExample("Bioelectrics/Monodomain")
+#buildExample("Bioelectrics/Monodomain")
 
-  buildExample("FluidMechanics/Stokes/ALE")
-  buildExample("FluidMechanics/Stokes/Static")
-  buildExample("FluidMechanics/Stokes/Dynamic")
+buildExample("FluidMechanics/Stokes/ALE")
+buildExample("FluidMechanics/Stokes/Static")
+buildExample("FluidMechanics/Stokes/Dynamic")
 
-  buildExample("FluidMechanics/NavierStokes/ALE")
-  buildExample("FluidMechanics/NavierStokes/Static")
-  buildExample("FluidMechanics/NavierStokes/Dynamic")
-
-
-  #buildExample("FluidMechanics/Darcy/ConvergenceStudy")
-  #buildExample("FluidMechanics/Darcy/FiveSpotProblem")
-  #buildExample("FluidMechanics/Darcy/VenousCompartment")
-
-  #buildExample("FiniteElasticity/UniAxialExtension")
-  #buildExample("FiniteElasticity/TwoElementTriLinear")
-  #buildExample("FiniteElasticity/MixedBoundaryConditions")
-  #buildExample("FiniteElasticity/TriCubicAxialExtension")
-
-  #buildExample("LinearElasticity/2DAnalytic1")
-  #buildExample("LinearElasticity/2DPlaneStressLagrangeBasis")
-  #buildExample("LinearElasticity/3DAnalytic1")
-  #buildExample("LinearElasticity/3DLagrangeBasis")
-  #buildExample("LinearElasticity/3DCubicHermiteBasis")
-  #buildExample("LinearElasticity/3DLagrangeBasisAnisotropicFibreField")
+buildExample("FluidMechanics/NavierStokes/ALE")
+buildExample("FluidMechanics/NavierStokes/Static")
+buildExample("FluidMechanics/NavierStokes/Dynamic")
 
 
-  # TODO Group them
-  #buildExample("LagrangeSimplexMesh")
-  #buildExample("cellml")
-  #buildExample("define-geometry-and-export")
-  #buildExample("MoreComplexMesh")
-  #buildExample("simple-field-manipulation-direct-access")
-  #buildExample("SimplexMesh")
-  #buildExample("TwoRegions")
+#buildExample("FluidMechanics/Darcy/ConvergenceStudy")
+#buildExample("FluidMechanics/Darcy/FiveSpotProblem")
+#buildExample("FluidMechanics/Darcy/VenousCompartment")
+
+#buildExample("FiniteElasticity/UniAxialExtension")
+#buildExample("FiniteElasticity/TwoElementTriLinear")
+#buildExample("FiniteElasticity/MixedBoundaryConditions")
+#buildExample("FiniteElasticity/TriCubicAxialExtension")
+
+#buildExample("LinearElasticity/2DAnalytic1")
+#buildExample("LinearElasticity/2DPlaneStressLagrangeBasis")
+#buildExample("LinearElasticity/3DAnalytic1")
+#buildExample("LinearElasticity/3DLagrangeBasis")
+#buildExample("LinearElasticity/3DCubicHermiteBasis")
+#buildExample("LinearElasticity/3DLagrangeBasisAnisotropicFibreField")
+
+
+# TODO Group them
+#buildExample("LagrangeSimplexMesh")
+#buildExample("cellml")
+#buildExample("define-geometry-and-export")
+#buildExample("MoreComplexMesh")
+#buildExample("simple-field-manipulation-direct-access")
+#buildExample("SimplexMesh")
+#buildExample("TwoRegions")
 
 f.close()
